+++ Content.Client/GameObjects/Components/Atmos/Piping/GasFilterVisualizer.cs
@@ -5,7 +5,6 @@ using Robust.Client.GameObjects;
 using Robust.Client.Interfaces.GameObjects.Components;
 using Robust.Shared.Interfaces.GameObjects;
 using Robust.Shared.Serialization;
-using Robust.Shared.Utility;
 using YamlDotNet.RepresentationModel;
 
 namespace Content.Client.GameObjects.Components.Atmos
