+++ Resources/Prototypes/Entities/Constructible/Ground/gascanisterports.yml
@@ -10,15 +10,12 @@
   - type: SnapGrid
     offset: Center
   - type: Sprite
+    netsync: false
     sprite: Constructible/Atmos/gascanisterport.rsi
-    state: gasCanisterPort
-  - type: Icon
-    sprite: Constructible/Atmos/gascanisterport.rsi
-    state: gasCanisterPort
-  - type: Appearance
-    visuals:
-    - type: PipeVisualizer
-      pipeRSI: Constructible/Atmos/pipe.rsi
+    layers:
+    - sprite: Constructible/Atmos/pipe.rsi
+      state: pipeHalf
+    - state: gasCanisterPort
   - type: Damageable
   - type: Destructible
     thresholds:
