+++ Resources/Prototypes/Entities/Constructible/Ground/pipes.yml
@@ -19,10 +19,10 @@
         - !type:DoActsBehavior
           acts: ["Destruction"]
   - type: Sprite
+    netsync: false
   - type: Appearance
     visuals:
     - type: PipeVisualizer
-      pipeRSI: Constructible/Atmos/pipe.rsi
   - type: Icon
     sprite: Constructible/Atmos/pipe.rsi
