--- Resources/Textures/Constructible/Atmos/pipeitems.rsi/meta.json
@@ -1,47 +1,0 @@
-{
-   "version":1,
-   "size":{
-      "x":32,
-      "y":32
-   },
-   "license":"CC-BY-SA-3.0",
-   "copyright":"Taken from https://github.com/tgstation/tgstation at commit 57cd1d59ca019dd0e7811ac451f295f818e573da",
-   "states":[
-     {
-       "name": "scrubber",
-       "directions": 4,
-       "delays": [
-         [
-           1
-         ],
-         [
-           1
-         ],
-         [
-           1
-         ],
-         [
-           1
-         ]
-       ]
-     },
-     {
-       "name": "vent",
-       "directions": 4,
-       "delays": [
-         [
-           1
-         ],
-         [
-           1
-         ],
-         [
-           1
-         ],
-         [
-           1
-         ]
-       ]
-     }
-   ]
-}
