+++ Resources/Textures/Constructible/Atmos/scrubber.rsi/meta.json
@@ -9,40 +9,12 @@
    "states":[
       {
          "name":"scrubOff",
-         "directions":1,
-         "delays":[ [ 1.0 ] ]
+         "directions":4
       },
       {
          "name":"scrubOn",
          "directions":1,
-         "delays":[
-            [
-               0.08,
-               0.08,
-               0.08,
-               0.08,
-               0.08,
-               0.08,
-               0.08,
-               0.08,
-               0.08,
-               0.08,
-               0.08,
-               0.08,
-               0.08,
-               0.08,
-               0.08,
-               0.08,
-               0.08,
-               0.08,
-               0.08,
-               0.08,
-               0.08,
-               0.08,
-               0.08,
-               0.08
-            ]
-         ]
+         "delays":[ [ 0.08, 0.08, 0.08, 0.08, 0.08, 0.08, 0.08, 0.08, 0.08, 0.08, 0.08, 0.08, 0.08, 0.08, 0.08, 0.08, 0.08, 0.08, 0.08, 0.08, 0.08, 0.08, 0.08, 0.08 ] ]
       }
    ]
 }
