+++ Resources/Textures/Constructible/Atmos/vent.rsi/meta.json
@@ -9,20 +9,12 @@
    "states":[
       {
          "name":"ventOff",
-         "directions":1,
-         "delays":[ [ 1.0 ] ]
+         "directions":4
       },
       {
          "name":"ventOn",
          "directions":1,
-         "delays":[
-            [
-               0.08,
-               0.08,
-               0.08,
-               0.08
-            ]
-         ]
+         "delays":[ [ 0.08, 0.08, 0.08, 0.08 ] ]
       }
    ]
 }
